requirements.txt
bs4
sentence_transformers
transformers
requests
numpy
pandas
flask
faiss-cpu